% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.test.R
\name{ADF.test}
\alias{ADF.test}
\title{ADF.test - simple implementation of ADF test}
\usage{
ADF.test(y, const = TRUE, trend = FALSE, max.lag = 0, criterion = "bic")
}
\arguments{
\item{y}{A variable of interest.}

\item{const}{Include const to the model if TRUE.}

\item{trend}{Include trend to the model if TRUE.}

\item{max.lag}{Maximum lag number}

\item{criterion}{A criterion used to select number of lags.}
}
\value{
List containing:
\itemize{
\item y
\item const
\item trend
\item residuals
\item coefficient estimates
\item t-statistic value
\item critical value
\item Number of lags
\item indicator of stationarity}
}
\description{
ADF.test - simple implementation of ADF test
}
