% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.test.S.R
\name{ADF.test.S}
\alias{ADF.test.S}
\title{Detrending bootstrap test by Smeekes (2013)}
\usage{
ADF.test.S(
  y,
  const = TRUE,
  trend = FALSE,
  c = 0,
  gamma = 0,
  trim = 0.15,
  max.lag = 0,
  criterion = NULL,
  modified.criterion = FALSE,
  iter = 999
)
}
\arguments{
\item{y}{The series of interest.}

\item{const}{Whether the constant is to included.}

\item{trend}{Whether the trend is to be included.}

\item{c}{A filtration parameter.}

\item{gamma}{A detrending type selection parameter. If 0 the OLS detrending
is applied, if 1 the GLS detrending is applied, otherwise the autocorrelation
coefficient is calculated as \eqn{1 + с^{\gamma} T^{-\gamma}}.}

\item{trim}{The trimming parameter.}

\item{max.lag}{The maximum lag for inner ADF testing.}

\item{criterion}{A criterion used to select number of lags.
If lag selection is not needed keep this NULL.}

\item{modified.criterion}{Whether the unit-root test modificaton is needed.}

\item{iter}{The number of bootstrap steps.}
}
\description{
Detrending bootstrap test by Smeekes (2013)
}
\details{
This bootstrap test is based on the recursive detrending procedure of Taylor
(2002). The main idea is to apply the standard ADF test to the series
with nuissanse parameters eliminated.

Critical values are calculated via a bootstrapping using MacKinnon-like
regressions. For each number of observations and each number of variables
obtained were 1999 values of test statistics. After that 1st, 2.5-th, 5-th,
10-th, and 97.5-th percentiles were calculated and saved along with the
corresponding number of observations. This step was repeated 5 times to cope
with possible biases. After that MacKunnon-like regressions were estimated.
}
\references{
Taylor, A. M. Robert.
“Regression-Based Unit Root Tests With Recursive Mean Adjustment for
Seasonal and Nonseasonal Time Series.”
Journal of Business & Economic Statistics 20, no. 2 (April 2002): 269–81.
https://doi.org/10.1198/073500102317352001.

MacKinnon, James G.
“Critical Values for Cointegration Tests.”
Working Paper. Economics Department, Queen’s University, January 2010.
https://ideas.repec.org/p/qed/wpaper/1227.html.

Smeekes, Stephan.
“Detrending Bootstrap Unit Root Tests.”
Econometric Reviews 32, no. 8 (July 2013): 869–91.
https://doi.org/10.1080/07474938.2012.690693.
}
