% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alrvr.R
\name{alrvr}
\alias{alrvr}
\title{Procedure ALRVR to estimate the long-run variance
as in Andrews (1991) and Kurozumi (2002).}
\usage{
alrvr(e)
}
\arguments{
\item{e}{(Tx1) vector or residuals.}
}
\value{
Long-run variance.
}
\description{
Procedure ALRVR to estimate the long-run variance
as in Andrews (1991) and Kurozumi (2002).
}
