% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alrvr_kernel.R
\name{alrvr_kernel}
\alias{alrvr_kernel}
\title{Calculating long-run variance}
\usage{
alrvr_kernel(e, max.lag = 0, kernel = "bartlett", criterion = "bic")
}
\arguments{
\item{e}{(Tx1) vector or residuals.}

\item{max.lag}{Maximum number of lags.
The exact number is selected by information criterions.}

\item{kernel}{Kernel for calculating long-run variance
\describe{
\item{bartlett}{for Bartlett kernel.}
\item{quadratic}{for Quadratic Spectral kernel.}
\item{NULL}{for the Kurozumi's proposal, using Bartlett kernel.}
}}

\item{criterion}{The information crietreion: bic, aic or lwz.}
}
\value{
Long-run variance.
}
\description{
Procedure ALRVR to estimate the long-run variance as
in Sul, Phillips and Choi (2003).
}
\details{
Used are Quadratic Spectral and Bartlett kernels.
}
