% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_kpss_mp.R
\name{bootstrap_kpss_mp}
\alias{bootstrap_kpss_mp}
\title{Calculating p-values using bootstrap}
\usage{
bootstrap_kpss_mp(
  y,
  x,
  model,
  break.point,
  const = FALSE,
  trend = FALSE,
  weakly.exog = TRUE,
  ll.init,
  corr.max,
  kernel,
  iter = 9999,
  bootstrap = "sample"
)
}
\arguments{
\item{y}{(Tx1)-vector of time series.}

\item{x}{(Txk)-matrix of explanatory stochastic regressors.}

\item{model}{A scalar or vector of
\describe{
\item{1}{for the break in const.}
\item{2}{for the break in trend.}
\item{3}{for the break in const and trend.}
}}

\item{break.point}{Array of structural breaks.}

\item{const}{Include constant if **TRUE**.}

\item{trend}{Include trend if **TRUE**.}

\item{weakly.exog}{Exogeneity of the stochastic regressors
\describe{
\item{TRUE}{if the regressors are weakly exogenous,}
\item{FALSE}{if the regressors are not weakly exogenous
(DOLS is used in this case).}
}}

\item{ll.init}{Scalar, defines the initial number of leads and lags for DOLS.}

\item{corr.max}{scalar, with the maximum order of the parametric correction.
The final order of the parametric correction is selected using the BIC
information criterion.}

\item{kernel}{Kernel for calculating long-run variance
\describe{
\item{bartlett}{for Bartlett kernel.}
\item{quadratic}{for Quadratic Spectral kernel.}
\item{NULL}{for the Kurozumi's proposal, using Bartlett kernel.}
}}

\item{iter}{Number of bootstrap iterations.}

\item{bootstrap}{Type of bootstrapping:
\describe{
\item{sample}{sampling from residuals with replacement.}
\item{Cavaliere-Taylor}{multiplying residuals by \eqn{N(0, 1)}-distributed
variable.}
\item{Rademacher}{multiplying residuals by Rademacher-distributed variable.}
}}
}
\value{
List of 3 elements:
\describe{
\item{test}{The value of KPSS test statistic.}
\item{p_value}{The estimates p-value.}
\item{bootstrapped}{Bootstrapped auxiliary statistics.}}
}
\description{
Procedure to compute the KPSS test with two structural breaks
}
\details{
The break points are known.

See Cavaliere and Taylor (2006) for further details.
}
