% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi_kpss.R
\name{coi_kpss}
\alias{coi_kpss}
\title{KPSS-test with known structural break}
\usage{
coi_kpss(y, x, model, tb, k2, cri)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable}

\item{x}{(Txk)-matrix of explanatory stochastic regressors}

\item{model}{(2x1)-vector, denotes the deterministic component:
\describe{
\item{model[1]}{\describe{
\item{1}{for model An}
\item{2}{for model A}
\item{3}{for model B}
\item{4}{for model C}
\item{5}{for model D}
\item{6}{for model E}}
}
\item{model[2]}{Position of the break point.}
}}

\item{k2}{Scalar. Indicates the deterministic component of the stochastic regressors:
\describe{
\item{1}{if x is a matrix of I(1) variables with drift.}
\item{2}{if x is a matrix of I(1) variables with a quadratic trend.}
}}

\item{cri}{(2x1)-vector where we specify whether the stochastic regressors are exogenous or not, and the initial number of leads and lags for the DOLS estimation.
\describe{
\item{cri[1]}{\describe{
\item{0}{if the regressors are weakly exogenous,}
\item{1}{if the regressors are not weakly exogenous (DOLS is used in this case).}
}
}
\item{cri[2]}{Scalar, defines the initial number of leads and lags for DOLS}
}}
}
\value{
\describe{
\item{beta}{DOLS estimates of the coefficients regressors.}
\item{tests}{SC test (coinkpss-test).}
\item{u}{Residuals of the model}
\item{t_b}{Individual significance t-statistics}
}
}
\description{
Computes the cointegration test with one known structural break.
}
\details{
The code provided is the original GAUSS code ported to R.

See Carrion-i-Silvestre and Sansó (2006) for further details.
}
