% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.test.R
\name{CPST.rescale}
\alias{CPST.rescale}
\title{Generating rescaled series as in Cavaliere et al. (2015).}
\usage{
CPST.rescale(d.y, x, deter, k, max.lag)
}
\arguments{
\item{d.y}{A series of first differences.}

\item{x}{A matrix of ADF RHS variables.}

\item{deter}{A matrix of deterministic variables for detrending.}

\item{k}{A lag of the corresponding ADF model.}

\item{max.lag}{The maximum possible lag.}
}
\value{
A rescaled series.
}
\description{
This rescaling procedure is needed to cope with possible heteroscedasticity
in the data. Simply it's achieved by taking a cumulative sum of the
first difference normalized by the non-parametric local estimate of the
variance.
}
\references{
Cavaliere, Giuseppe, Peter C. B. Phillips, Stephan Smeekes,
and A. M. Robert Taylor. “Lag Length Selection for Unit Root Tests
in the Presence of Nonstationary Volatility.”
Econometric Reviews 34, no. 4 (April 21, 2015): 512–36.
https://doi.org/10.1080/07474938.2013.808065.
}
