% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.values.R
\name{critical.values.break.date.cset}
\alias{critical.values.break.date.cset}
\title{Critical values for break point confidence intervals}
\usage{
critical.values.break.date.cset(l_1, trend, conf.level, p_zb, p_zf)
}
\arguments{
\item{l_1}{Relative break point position.}

\item{trend}{Whether thern is to be included.}

\item{conf.level}{Confidense level.}

\item{p_zb}{Number of variables with breaks.}

\item{p_zf}{Number of variables without breaks.}
}
\description{
Auxiliary function returning pre-calculated critical values for
\link{break.date.cset}
}
\details{
The function is not intended to be used directly so it's not exported.
}
