% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.values.R
\name{critical.values.KPSS.2p}
\alias{critical.values.KPSS.2p}
\title{Auxiliary function returning pre-calculated critical values for
[KPSS.2.breaks()].}
\usage{
critical.values.KPSS.2p(model, break.point, N)
}
\arguments{
\item{model}{\describe{
\item{1}{for the AA (without trend) model.}
\item{2}{for the AA (with trend) model.}
\item{3}{for the BB model.}
\item{4}{for the CC model.}
\item{5}{for the AC-CA model.}
}}

\item{break.point}{Position of the break point.}

\item{N}{Number of observations.}
}
\description{
Auxiliary function returning pre-calculated critical values for
[KPSS.2.breaks()].
}
\details{
The function is not intended to be used directly so it's not exported.
}
