% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determi_kpss_1p.R
\name{determi_kpss_1p}
\alias{determi_kpss_1p}
\title{Construct determinant variables.}
\usage{
determi_kpss_1p(model, N, break.point)
}
\arguments{
\item{model}{\describe{
\item{1}{Model with trend, break in const.}
\item{2}{Model with const and trend, break in const.}
\item{3}{Model with const and trend, break in trend.}
\item{4}{Model with const and trend, break in const and trend.}
}}

\item{N}{Number of observations.}

\item{break.point}{Break point.}
}
\value{
Matrix of determinant variables.
}
\description{
Construct determinant variables.
}
