% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determi_kpss_mp.R
\name{determi_kpss_mp}
\alias{determi_kpss_mp}
\title{Procedure to compute deterministic terms
for KPSS with 2 structural breaks.}
\usage{
determi_kpss_mp(model, N, break_point, trend = FALSE)
}
\arguments{
\item{model}{A scalar or vector of
\describe{
\item{1}{for the break in const.}
\item{2}{for the break in trend.}
\item{3}{for the break in const and trend.}
}}

\item{N}{Number of observations.}

\item{break_point}{Array of structural breaks.}

\item{trend}{Include trend if `TRUE`.}
}
\value{
Matrix of deterministic terms.
}
\description{
Procedure to compute deterministic terms
for KPSS with 2 structural breaks.
}
