% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinants.R
\name{determinants.KPSS.N.breaks}
\alias{determinants.KPSS.N.breaks}
\title{Deterministic terms for \link{KPSS.N.breaks}}
\usage{
determinants.KPSS.N.breaks(model, N, break.point, const = FALSE, trend = FALSE)
}
\arguments{
\item{model}{A scalar or vector of
\itemize{
\item 1: for the break in const,
\item 2: for the break in trend,
\item 3: for the break in const and trend.
}}

\item{N}{Number of observations.}

\item{break.point}{Array of structural breaks.}

\item{const, trend}{Include constant and trend if \code{TRUE}.}
}
\value{
Matrix of deterministic terms.
}
\description{
Procedure to compute deterministic terms for KPSS with \eqn{m}
structural breaks.
}
\details{
\strong{model} should be either a scalar or a vector of the same size as the
\strong{break.point}. If scalar \strong{model} will be repeated till the length of
\strong{break.point} is achieved.
}
