% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.N.breaks.R
\name{DOLS.N.breaks}
\alias{DOLS.N.breaks}
\title{Estimating DOLS regression for multiple known break points}
\usage{
DOLS.N.breaks(
  y,
  x,
  model,
  break.point,
  const = FALSE,
  trend = FALSE,
  k.lags,
  k.leads
)
}
\arguments{
\item{y}{A dependent (LHS) variable.}

\item{x}{A matrix of explanatory (RHS) variables.}

\item{model}{A scalar or vector of break types:
\itemize{
\item 1: for the break in const.
\item 2: for the break in trend.
\item 3: for the break in const and trend.
}}

\item{break.point}{An array of moments of structural breaks.}

\item{const, trend}{Whether a constant or trend are to be included.}

\item{k.lags, k.leads}{A number of lags and leads in DOLS regression.}
}
\value{
A list of:
\itemize{
\item Estimates of coefficients,
\item Estimates of residuals,
\item A set of informational criterions values,
\item \eqn{t}-statistics for the estimates of coefficients.
}
}
\description{
Estimating DOLS regression for multiple known break points
}
\details{
The function is not intended to be used directly so it's not exported.
}
