% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOLS.R
\name{DOLS}
\alias{DOLS}
\title{Estimating DOLS regression for multiple known break points}
\usage{
DOLS(y, x, model, break.point, k.lags, k.leads)
}
\arguments{
\item{y}{A dependent (LHS) variable.}

\item{x}{A matrix of explanatory (RHS) variables.}

\item{model}{See Carrion-i-Silvestre and Sansó (2006)
\itemize{
\item 1: for model An,
\item 2: for model A,
\item 3: for model B,
\item 4: for model C,
\item 5: for model D,
\item 6: for model E.
}}

\item{break.point}{A position of the break point.}

\item{k.lags, k.leads}{A number of lags and leads in DOLS regression.}
}
\value{
A list of:
\itemize{
\item Estimates of coefficients,
\item Estimates of residuals,
\item A value of BIC,
\item \eqn{t}-statistics for the estimates of coefficients.
}
}
\description{
Estimating DOLS regression for multiple known break points
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Carrion-i-Silvestre, Josep Lluís, and Andreu Sansó.
“Testing the Null of Cointegration with Structural Breaks.”
Oxford Bulletin of Economics and Statistics 68, no. 5 (October 2006): 623–46.
https://doi.org/10.1111/j.1468-0084.2006.00180.x.
}
