% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.values.R
\name{get.cv.KPSS.1p}
\alias{get.cv.KPSS.1p}
\title{Critical values for KPSS test with 1 break}
\usage{
get.cv.KPSS.1p(model, break.point, N, k)
}
\arguments{
\item{model}{A scalar equal to
\itemize{
\item 1: for model An,
\item 2: for model A,
\item 3: for model B,
\item 4: for model C,
\item 5: for model D,
\item 6: for model E.
}}

\item{break.point}{Position of the break point.}

\item{N}{Number of observations.}

\item{k}{Number of RHS variables.}
}
\description{
Auxiliary function returning pre-calculated critical values for
\link{KPSS.1.break}
}
\details{
The function is not intended to be used directly so it's not exported.
}
