% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.values.R
\name{get.cv.KPSS.2p}
\alias{get.cv.KPSS.2p}
\title{Critical values for KPSS test with 2 breaks}
\usage{
get.cv.KPSS.2p(model, break.point, N)
}
\arguments{
\item{model}{A scalar equal to
\itemize{
\item 1: for the AA (without trend) model,
\item 2: for the AA (with trend) model,
\item 3: for the BB model,
\item 4: for the CC model,
\item 5: for the AC-CA model.
}}

\item{break.point}{Position of the break point.}

\item{N}{Number of observations.}
}
\description{
Auxiliary function returning pre-calculated critical values for
\link{KPSS.2.breaks}
}
\details{
The function is not intended to be used directly so it's not exported.
}
