% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDF.CHLT.R
\name{GLS.bt}
\alias{GLS.bt}
\title{GLS fitering}
\usage{
GLS.bt(y, trim, c)
}
\arguments{
\item{y}{Series of interest.}

\item{trim}{Trimming parameter.}

\item{c}{Filtering parameter.}
}
\description{
GLS fitering
}
