% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS-GLS.R
\name{GLS}
\alias{GLS}
\title{Custom GLS with extra information}
\usage{
GLS(y, z, c)
}
\arguments{
\item{y}{Dependent variable.}

\item{c}{Coefficient for \eqn{\rho} calculation.}

\item{x}{Explanatory variables.}
}
\value{
The list of betas, residuals, forecasted values and t-values.
}
\description{
Getting GLS estimates of betas, residuals, forecasted values and t-values.
}
