% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSADF.bootstrap.test.R
\name{GSADF.bootstrap.test}
\alias{GSADF.bootstrap.test}
\title{SADF test with wild bootstrap.}
\usage{
GSADF.bootstrap.test(
  y,
  r0 = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  alpha = 0.05,
  iter = 4 * 200,
  seed = round(10^4 * sd(y))
)
}
\description{
SADF test with wild bootstrap.
}
