% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSADF.test.R
\name{GSADF.test}
\alias{GSADF.test}
\title{Generalized supremum ADF test.}
\usage{
GSADF.test(
  y,
  trim = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  add.p.value = TRUE
)
}
\arguments{
\item{y}{The input time series of interest.}

\item{trim}{Trimming parameter to determine the lower and upper bounds.}

\item{const}{Whether the constant needs to be included.}

\item{add.p.value}{Whether the p-value is to be returned. This argument is
needed to suppress the calculation of p-values during the precalculation of
tables needed for the p-values estimating.}
}
\description{
Generalized supremum ADF test.
}
\references{
Kurozumi, Eiji, Anton Skrobotov, and Alexey Tsarev.
“Time-Transformed Test for Bubbles under Non-Stationary Volatility.”
Journal of Financial Econometrics, April 23, 2022.
https://doi.org/10.1093/jjfinec/nbac004.
}
