% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSTADF.test.R
\name{GSTADF.test}
\alias{GSTADF.test}
\title{GSADF test based on NW estimation and reindexing.}
\usage{
GSTADF.test(
  y,
  r0 = 0.01 + 1.8/sqrt(length(y)),
  const = FALSE,
  omega.est = TRUE,
  h = "auto_CV",
  truncated = TRUE,
  is.reindex = TRUE,
  ksi.input = "auto",
  hc = 1,
  pc = 1,
  add.p.value = TRUE
)
}
\description{
GSADF test based on NW estimation and reindexing.
}
