% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.criterion.R
\name{info.criterion}
\alias{info.criterion}
\title{Information criterions}
\usage{
info.criterion(resid, extra, modification = FALSE, alpha = 0, y = NULL)
}
\arguments{
\item{resid}{Input residuals.}

\item{extra}{Number of coefficients etc.}

\item{modification}{Whether the unit-root test modificaton is needed.}

\item{alpha}{The coefficient \eqn{\alpha} of \eqn{y_{t-1}} in ADF model.}

\item{y}{The vector of \eqn{y_{t-1}} in ADF model.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}
}
\value{
The value of the selected informational criterion.
}
\description{
Information criterions
}
\details{
Calculating the value of the following informational criterions:
\itemize{
\item Akaike,
\item Schwarz (Bayesian),
\item Hannan-Quinn,
\item Liu et al.}
}
