% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KP.R
\name{KP}
\alias{KP}
\title{Kejrival-Perron procedure of breaks number detection.}
\usage{
KP(y, const = FALSE, breaks = 1, criterion = "aic", trim = 0.15)
}
\arguments{
\item{y}{The input series of interest.}

\item{const}{Allowing the break in constant.}

\item{breaks}{Number of breaks.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}

\item{trim}{A trimming value for a possible break date bounds.}
}
\description{
Kejrival-Perron procedure of breaks number detection.
}
