% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.N.breaks.bootstrap.R
\name{KPSS.N.breaks.bootstrap}
\alias{KPSS.N.breaks.bootstrap}
\title{KPSS-test with multiple unknown structural breaks}
\usage{
KPSS.N.breaks.bootstrap(
  y,
  x,
  model,
  break.point,
  const = FALSE,
  trend = FALSE,
  weakly.exog = TRUE,
  lags.init,
  leads.init,
  max.lag,
  kernel,
  iter = 9999,
  bootstrap = "sample",
  criterion = "bic"
)
}
\arguments{
\item{y}{An input (LHS) time series of interest.}

\item{x}{A matrix of (RHS) explanatory stochastic regressors.}

\item{model}{A scalar or vector of
\itemize{
\item 1: for the break in const,
\item 2: for the break in trend,
\item 3: for the break in const and trend.
}}

\item{break.point}{Array of structural breaks.}

\item{const}{Include constant if \strong{TRUE}.}

\item{trend}{Include trend if \strong{TRUE}.}

\item{weakly.exog}{Boolean where we specify
whether the stochastic regressors are exogenous or not
\itemize{
\item \code{TRUE}: if the regressors are weakly exogenous,
\item \code{FALSE}: if the regressors are not weakly exogenous
(DOLS is used in this case).
}}

\item{lags.init, leads.init}{Scalars defininig the initial number of lags and
leads for DOLS.}

\item{max.lag}{scalar, with the maximum order of the parametric correction.
The final order of the parametric correction is selected using the BIC
information criterion.}

\item{kernel}{Kernel for calculating long-run variance
\itemize{
\item \code{bartlett}: for Bartlett kernel,
\item \code{quadratic}: for Quadratic Spectral kernel,
\item \code{NULL} for the Kurozumi's proposal, using Bartlett kernel.
}}

\item{iter}{Number of bootstrap iterations.}

\item{bootstrap}{Type of bootstrapping:
\itemize{
\item \code{"sample"}: sampling from residuals with replacement,
\item \code{"Cavaliere-Taylor"}: multiplying residuals by \eqn{N(0, 1)}-distributed
variable,
\item \code{"Rademacher"}: multiplying residuals by Rademacher-distributed variable.
}}

\item{criterion}{Information criterion for DOLS lags and leads selection:
aic, bic or lwz.}
}
\value{
A list of:
\itemize{
\item \code{test}: The value of KPSS test statistic,
\item \code{p.value}: The estimates p-value,
\item \code{bootstrapped}: Bootstrapped auxiliary statistics.
}
}
\description{
Procedure to compute the KPSS test with multiple unknown structural breaks
}
