% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss_known_2p.R
\name{kpss_known_2p}
\alias{kpss_known_2p}
\title{KPSS-test with 2 known structural breaks}
\usage{
kpss_known_2p(y, model, break.point, kmax, kernel)
}
\arguments{
\item{y}{(Tx1)-vector of time series}

\item{model}{\describe{
\item{1}{for the AA (without trend) model.}
\item{2}{for the AA (with trend) model.}
\item{3}{for the BB model.}
\item{4}{for the CC model.}
\item{5}{for the AC-CA model.}
}}

\item{kmax}{scalar, with the maximum order of the parametric correction.
The final order of the parametric correction is selected using the
BIC information criterion.}

\item{kernel}{Kernel for calculating long-run variance
\describe{
\item{bartlett}{for Bartlett kernel.}
\item{quadratic}{for Quadratic Spectral kernel.}
\item{NULL}{for the Kurozumi's proposal, using Bartlett kernel.}
}}

\item{tb1}{The first break point.}

\item{tb2}{The second break point.}
}
\value{
\describe{
\item{beta}{DOLS estimates of the coefficients regressors.}
\item{tests}{SC test (coinkpss-test).}
\item{resid}{Residuals of the model.}
\item{break_point}{Break points.}
}
}
\description{
Procedure to compute the KPSS test with two structural breaks
}
\details{
The break points are known

The code provided is the original GAUSS code ported to R.

See Carrion-i-Silvestre and Sansó (2007) for further details.
}
