% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss_known_mp.R
\name{kpss_known_mp}
\alias{kpss_known_mp}
\title{KPSS-test with 2 known structural breaks}
\usage{
kpss_known_mp(
  y,
  x,
  model,
  break_point,
  trend = FALSE,
  weakly.exog = TRUE,
  kinit,
  kmax,
  kernel
)
}
\arguments{
\item{y}{(Tx1)-vector of time series.}

\item{x}{(Txk)-matrix of explanatory stochastic regressors.}

\item{model}{A scalar or vector of
\describe{
\item{1}{for the break in const.}
\item{2}{for the break in trend.}
\item{3}{for the break in const and trend.}
}}

\item{break_point}{Array of structural breaks.}

\item{trend}{Include trend if `TRUE`.}

\item{weakly.exog}{Boolean where we specify whether the stochastic regressors are exogenous or not
\describe{
\item{TRUE}{if the regressors are weakly exogenous,}
\item{FALSE}{if the regressors are not weakly exogenous (DOLS is used in this case).}
}}

\item{kinit}{Scalar, defines the initial number of leads and lags for DOLS.}

\item{kmax}{scalar, with the maximum order of the parametric correction. The final order of the parametric correction is selected using the BIC information criterion.}

\item{kernel}{\describe{
\item{bartlett}{for Bartlett kernel.}
\item{quadratic}{for Quadratic Spectral kernel.}
\item{NULL}{for the Kurozumi's proposal, using Bartlett kernel.}
}}
}
\value{
\describe{
\item{beta}{DOLS estimates of the coefficients regressors.}
\item{tests}{SC test (coinkpss-test).}
\item{resid}{Residuals of the model.}
\item{break_point}{Break points.}
}
}
\description{
Procedure to compute the KPSS test with two structural breaks
}
\details{
The break points are known

The code provided is the original GAUSS code ported to R.

See Carrion-i-Silvestre and Sansó (2006) for further details.
}
