% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagn.R
\name{lagn}
\alias{lagn}
\title{Produce lagged vector backward of forward}
\usage{
lagn(x, i, na = NA)
}
\arguments{
\item{x}{Initial vector.}

\item{i}{Size of lag (lead if negative).}
}
\value{
Lagged or leaded vector.
}
\description{
Produce lagged vector backward of forward
}
