% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.var.R
\name{lr.var.AK}
\alias{lr.var.AK}
\title{Calculating long-run variance}
\usage{
lr.var.AK(e)
}
\arguments{
\item{e}{(Tx1) vector or residuals.}
}
\value{
Long-run variance.
}
\description{
Procedure ALRVR to estimate the long-run variance
as in Andrews (1991) and Kurozumi (2002).
}
