% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.var.R
\name{lr.var.bartlett}
\alias{lr.var.bartlett}
\title{Calculating long-run variance}
\usage{
lr.var.bartlett(e, l = NULL)
}
\arguments{
\item{e}{(Tx1) vector or residuals.}
}
\value{
Long-run variance.
}
\description{
Calculating long-run variance
}
