% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.var.R
\name{lr.var.quad}
\alias{lr.var.quad}
\title{Estimating heteroscedasticity and autocorrelation consistent variance}
\usage{
lr.var.quad(y)
}
\description{
Estimating heteroscedasticity and autocorrelation consistent variance
}
\references{
Andrews, Donald W. K.
“Heteroskedasticity and Autocorrelation Consistent
Covariance Matrix Estimation.”
Econometrica 59, no. 3 (1991): 817–58.
https://doi.org/10.2307/2938229.
}
