% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.var.merged.R
\name{lr.var}
\alias{lr.var}
\title{TEST}
\usage{
lr.var(
  y,
  kernel = "bartlett",
  limit.lags = FALSE,
  limit.selector = "kpss-q",
  upper.rho.limit = 0.97,
  upper.lag.limit = 0.8,
  recolor = FALSE,
  max.lag = 0,
  criterion = "bic"
)
}
\description{
TEST
}
