% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDF.CHLT.R
\name{MDF.CHLT}
\alias{MDF.CHLT}
\title{MDF test for a single break and possible heterscedasticity}
\usage{
MDF.CHLT(y, max.lag = 10, trim = 0.15, iter = 499)
}
\arguments{
\item{y}{The series of interest}

\item{max.lag}{The maximum possible lag.}

\item{trim}{The trimming parameter for lag selection}
}
\description{
MDF test for a single break and possible heterscedasticity
}
\references{
Cavaliere, Giuseppe, David I. Harvey, Stephen J. Leybourne,
and A.M. Robert Taylor.
“Testing for Unit Roots in the Presence of a Possible Break in Trend and
Nonstationary Volatility.”
Econometric Theory 27, no. 5 (October 2011): 957–91.
https://doi.org/10.1017/S0266466610000605.
}
