% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW.estimation.R
\name{NW.estimation}
\alias{NW.estimation}
\title{NW.estimation - Nadaraya–Watson kernel regression.}
\usage{
NW.estimation(y, x, h, kernel = "unif")
}
\arguments{
\item{y}{LHS dependent variable.}

\item{x}{RHS explanation variable.}

\item{h}{Bandwidth.}

\item{kernel}{Needed kernel, currently only `unif` and `gauss`.}
}
\value{
A list of arguments as well as the estimated coefficient vector and
residuals.
}
\description{
NW.estimation - Nadaraya–Watson kernel regression.
}
\details{
See Harvey, Leybourne and Zu (2021) for more information.
}
