% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.values.SADF.R
\name{p.values.SADF}
\alias{p.values.SADF}
\title{Interpolating p-value for intermediate observation numbers.}
\usage{
p.values.SADF(statistic, N.obs, cr.values)
}
\arguments{
\item{statistic}{The statistic value.}

\item{N.obs}{The number of observations.}

\item{cr.values}{The set of precalculated tables.}
}
\description{
Interpolating p-value for intermediate observation numbers.
}
