% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.results.R
\name{print.sadf}
\alias{print.sadf}
\alias{print.mdfHLT}
\alias{print.mdfHLTN}
\alias{print.mdfHHLT}
\alias{print.cointGH}
\title{Custom functions for printing results in a nice way.}
\usage{
\method{print}{sadf}(x)

\method{print}{mdfHLT}(x)

\method{print}{mdfHLTN}(x)

\method{print}{mdfHHLT}(x)

\method{print}{cointGH}(x)
}
\arguments{
\item{x}{Object containing results.}
}
\description{
Custom functions for printing results in a nice way.
}
