% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PY.R
\name{PY}
\alias{PY}
\title{Perron-Yabu (2009) statistic for break at unknown date.}
\usage{
PY(y, const = FALSE, trend = FALSE, criterion = "aic", trim = 0.15, max.lag)
}
\arguments{
\item{y}{The input series of interest.}

\item{const}{Allowing the break in constant.}

\item{trend}{Allowing the break in trend.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}

\item{trim}{A trimming value for a possible break date bounds.}

\item{max.lag}{The maximum possible lag in the model.}
}
\description{
Perron-Yabu (2009) statistic for break at unknown date.
}
