% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PY.sequential.R
\name{PY.sequential}
\alias{PY.sequential}
\title{Sequential Perron-Yabu (2009) statistic for breaks at unknown date.}
\usage{
PY.sequential(
  y,
  const = FALSE,
  breaks = 1,
  criterion = "aic",
  trim = 0.15,
  max.lag = 1
)
}
\arguments{
\item{y}{The input series of interest.}

\item{const}{Allowing the break in constant.}

\item{breaks}{Number of breaks.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}

\item{trim}{A trimming value for a possible break date bounds.}

\item{max.lag}{The maximum possible lag in the model.}
}
\description{
Sequential Perron-Yabu (2009) statistic for breaks at unknown date.
}
