% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PY.single.R
\name{PY.single}
\alias{PY.single}
\title{Perron-Yabu (2009) statistic for break at unknown date.}
\usage{
PY.single(
  y,
  const = FALSE,
  trend = FALSE,
  criterion = "aic",
  trim = 0.15,
  max.lag
)
}
\arguments{
\item{y}{The input series of interest.}

\item{const, trend}{Allowing the break in constant or trend.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}

\item{trim}{A trimming value for a possible break date bounds.}

\item{max.lag}{The maximum possible lag in the model.}
}
\value{
A list of the estimated Wald statistic as well as its c.v.
}
\description{
Perron-Yabu (2009) statistic for break at unknown date.
}
\references{
Perron, Pierre, and Tomoyoshi Yabu.
“Testing for Shifts in Trend With an Integrated or
Stationary Noise Component.”
Journal of Business & Economic Statistics 27, no. 3 (July 2009): 369–96.
https://doi.org/10.1198/jbes.2009.07268.
}
