% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindex.R
\name{reindex}
\alias{reindex}
\title{reindex - function that makes reindexing.}
\usage{
reindex(u)
}
\description{
reindex - function that makes reindexing.
}
\details{
See Cavaliere and Taylor (2007, 2008) and
Kurozumi, Skrobotov and Tsaryov (2021) for more information.
}
