% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.tests.R
\name{robust.tests.multiple}
\alias{robust.tests.multiple}
\title{A wrapping function around \code{\link[=KP]{KP()}} and \code{\link[=MDF.multiple]{MDF.multiple()}}.}
\usage{
robust.tests.multiple(
  y,
  const = FALSE,
  season = FALSE,
  breaks = 2,
  trim = 0.15
)
}
\arguments{
\item{y}{A series of interest.}

\item{const}{Whether the constant term should be included.}

\item{season}{Whether the seasonal adjustment is needed.}

\item{breaks}{Number of breaks.}

\item{trim}{Trimming value for a possible break date bounds.}
}
\description{
A wrapping function around \code{\link[=KP]{KP()}} and \code{\link[=MDF.multiple]{MDF.multiple()}}.
}
