% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb.GSADF.test.R
\name{sb.GSADF.test}
\alias{sb.GSADF.test}
\title{Sign-based SADF test (HLZ, 2019).}
\usage{
sb.GSADF.test(
  y,
  r0 = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  alpha = 0.05,
  iter = 4 * 200,
  urs = TRUE,
  seed = round(10^4 * sd(y))
)
}
\description{
Sign-based SADF test (HLZ, 2019).
}
