% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinants.R
\name{seasonal.dummies}
\alias{seasonal.dummies}
\title{Generating monthly seasonal dummy variables}
\usage{
seasonal.dummies(N)
}
\arguments{
\item{N}{number of observations.}
}
\value{
The matrix of values od seasonal dummies.
}
\description{
Generating monthly seasonal dummy variables
}
\details{
The function is not intended to be used directly so it's not exported.
}
