% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.points.R
\name{segments.GLS}
\alias{segments.GLS}
\title{Procedure to minimize the GLS-SSR for 1 break point}
\usage{
segments.GLS(
  y,
  const = FALSE,
  trend = FALSE,
  breaks = 1,
  first.break = NULL,
  last.break = NULL,
  trim = 0.15
)
}
\arguments{
\item{y}{Variable of interest.}

\item{const}{Whether there is a break in the constant.}

\item{trend}{Whether there is a break in the trend.}

\item{breaks}{Number of breaks.}

\item{first.break}{First possible break point.}

\item{last.break}{Last possible break point.}

\item{trim}{Trim value to calculate `first.break` and `last.break`
if not provided.}
}
\value{
The point of possible break.
}
\description{
Procedure to minimize the GLS-SSR for 1 break point
}
\details{
See Skrobotov (2018) for further details.
}
