% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.Segments.R
\name{segments.OLS.1.break}
\alias{segments.OLS.1.break}
\title{Procedure to minimize the SSR for 1 break point}
\usage{
segments.OLS.1.break(beg, end, first.break, last.break, len, SSR.data)
}
\arguments{
\item{beg}{Start of the sample.}

\item{end}{End of the sample.}

\item{first.break}{First possible break point.}

\item{last.break}{Last possible break point.}

\item{len}{Total number of observations.}

\item{SSR.data}{The matrix of recursive SSR values.}
}
\value{
A list of:
\itemize{
\item \code{SSR}: Optimal SSR value,
\item \code{break.point}: The point of possible break.
}
}
\description{
Procedure to minimize the SSR for 1 break point
}
\references{
Carrion-i-Silvestre, Josep Lluís, and Andreu Sansó.
“Testing the Null of Cointegration with Structural Breaks.”
Oxford Bulletin of Economics and Statistics 68, no. 5 (October 2006): 623–46.
https://doi.org/10.1111/j.1468-0084.2006.00180.x.
}
\keyword{internal}
