% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.Segments.R
\name{segments.OLS.N.breaks}
\alias{segments.OLS.N.breaks}
\title{Find \eqn{m + 1} optimal partitions}
\usage{
segments.OLS.N.breaks(y, x, m = 1, width = 2, SSR.data = NULL)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable.}

\item{x}{(Txk)-vector of the explanatory stochastic regressors.}

\item{m}{Number of breaks.}

\item{width}{Minimum spacing between the breaks.}

\item{SSR.data}{Optional matrix of recursive SSR's.}
}
\value{
A list of:
\itemize{
\item optimal SSR,
\item the vector of break points.
}
}
\description{
Find \eqn{m + 1} optimal partitions
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Bai, Jushan, and Pierre Perron.
“Computation and Analysis of Multiple Structural Change Models.”
Journal of Applied Econometrics 18, no. 1 (2003): 1–22.
https://doi.org/10.1002/jae.659.
}
