% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.OLS.R
\name{segments.OLS}
\alias{segments.OLS}
\title{Find \eqn{m + 1} optimal partitions}
\usage{
segments.OLS(y, x, m = 1, width = 2, SSR.data = NULL)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable.}

\item{x}{(Txk)-vector of the explanatory stochastic regressors.}

\item{m}{Number of breaks.}

\item{width}{Minimum spacing between the breaks.}

\item{SSR.data}{Optional matrix of recursive SSR's.}
}
\value{
List of 2 elements: optimal SSR and the vector of break points.
}
\description{
Find \eqn{m + 1} optimal partitions
}
\references{
Bai, Jushan, and Pierre Perron.
“Computation and Analysis of Multiple Structural Change Models.”
Journal of Applied Econometrics 18, no. 1 (2003): 1–22.
https://doi.org/10.1002/jae.659.
}
