% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.OLS.single.R
\name{segments.OLS.single}
\alias{segments.OLS.single}
\title{Procedure to minimize the SSR for 1 break point}
\usage{
segments.OLS.single(beg, end, first.break, last.break, len, SSR.data)
}
\arguments{
\item{beg}{Sample begin.}

\item{end}{Sample end.}

\item{first.break}{First possible break point.}

\item{last.break}{Last possible break point.}

\item{len}{Total number of observations.}

\item{SSR.data}{The matrix of recursive SSR values.}
}
\value{
List containing
\describe{
\item{SSR}{Optimal SSR value.}
\item{break_point}{The point of possible break.}
}
}
\description{
Procedure to minimize the SSR for 1 break point
}
\details{
See Carrion-i-Silvestre and Sansó (2006) for further details.
}
\references{
Carrion-i-Silvestre, Josep Lluís, and Andreu Sansó.
“Testing the Null of Cointegration with Structural Breaks.”
Oxford Bulletin of Economics and Statistics 68, no. 5 (October 2006): 623–46.
https://doi.org/10.1111/j.1468-0084.2006.00180.x.
}
