% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segs.ssr.1.break.R
\name{segs.ssr.1.break}
\alias{segs.ssr.1.break}
\title{Procedure to minimize the SSR for 1 break point}
\usage{
segs.ssr.1.break(beg, end, first.break, last.break, len, ssr.data)
}
\arguments{
\item{beg}{Sample begin.}

\item{end}{Sample end.}

\item{first.break}{First possible break point.}

\item{last.break}{Last possible break point.}

\item{len}{Total number of observations.}

\item{ssr.data}{The matrix of recursive SSR values.}
}
\value{
List containing
\describe{
\item{ssr}{Optimal SSR value.}
\item{break_point}{The breaking point.}
}
}
\description{
Procedure to minimize the SSR for 1 break point
}
\details{
See Carrion-i-Silvestre and Sansó (2006) for further details.
}
