% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coint.conf.sets.R
\name{select.lead.lag}
\alias{select.lead.lag}
\title{Estimating optimal number of leads and lags}
\usage{
select.lead.lag(
  y,
  trend = TRUE,
  zb = NULL,
  zf = NULL,
  trim = 0.05,
  criterion = "bic"
)
}
\arguments{
\item{y}{LHS dependent variable.}

\item{trend}{Whether the trend is to be included.}

\item{zb}{I(1) regressors with break.}

\item{zf}{I(1) regressors without break.}

\item{trim}{The trimming parameter to find the lower and upper bounds of
possible break date.}

\item{criterion}{A criterion for lead and lag number estimation.}
}
\value{
A list of estimated values of leads and lags.
}
\description{
Estimating optimal number of leads and lags
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Kurozumi, Eiji, and Anton Skrobotov.
“Confidence Sets for the Break Date in Cointegrating Regressions.”
Oxford Bulletin of Economics and Statistics 80, no. 3 (2018): 514–35.
https://doi.org/10.1111/obes.12223.
}
