% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr_partition_1p.R
\name{ssr_partition_1p}
\alias{ssr_partition_1p}
\title{Procedure to minimize the SSR for 2 break points}
\usage{
ssr_partition_1p(beg, end, first_b, last_b, len, ssr_data)
}
\arguments{
\item{beg}{Sample begin.}

\item{end}{Sample end.}

\item{first_b}{First possible break point.}

\item{last_b}{Last possible break point.}

\item{len}{Total number of observations.}

\item{ssr_data}{The matrix of recursive SSR values.}
}
\value{
List containing
\describe{
\item{ssr}{Optimal SSR value.}
\item{break_point}{The breaking point.}
}
}
\description{
Procedure to minimize the SSR for 2 break points
}
