% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr_partition_2p.R
\name{ssr_partition_2p}
\alias{ssr_partition_2p}
\title{Procedure to minimize the SSR for 2 break points}
\usage{
ssr_partition_2p(y, model)
}
\arguments{
\item{y}{(Tx1)-vector of time series}

\item{model}{\describe{
\item{1}{for the AA (without trend) model.}
\item{2}{for the AA (with trend) model.}
\item{3}{for the BB model.}
\item{4}{for the CC model.}
\item{5}{for the AC-CA model.}
}}
}
\value{
List containing
\describe{
\item{resid}{(Tx1) vector of estimated OLS residuals.}
\item{tb1}{The first breaking point.}
\item{tb2}{The second breaking point.}
}
}
\description{
Procedure to minimize the SSR for 2 break points
}
