% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr_partition_mp.R
\name{ssr_partition_mp}
\alias{ssr_partition_mp}
\title{Find m + 1 optimal partitions}
\usage{
ssr_partition_mp(y, x, m = 1, width = 2, ssr_data = NULL)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable.}

\item{x}{(Txk)-vector of the explanatory stochastic regressors.}

\item{m}{Number of breaks.}

\item{width}{Minimum spacing between the breaks.}
}
\value{
List of 2 elements: optimal SSR and the vector of breaks.
}
\description{
Find m + 1 optimal partitions
}
\details{
Based on Bai and Perron (2003).
}
