% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr_recursive.R
\name{ssr_recursive}
\alias{ssr_recursive}
\title{Calculate SSR recursively}
\usage{
ssr_recursive(y, x, beg, end, width = 2)
}
\description{
Calculate SSR recursively
}
\details{
Based on Brown, Durbin and Evans (1975).
}
