% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supBZ.statistic.R
\name{supBZ.statistic}
\alias{supBZ.statistic}
\title{supBZ.statistic - Calculate supBZ statistic (HLZ, 2018).}
\usage{
supBZ.statistic(
  y,
  r0 = 0.01 + 1.8/sqrt(length(y)),
  sigma.sq = "auto",
  generalized = FALSE
)
}
\description{
supBZ.statistic - Calculate supBZ statistic (HLZ, 2018).
}
