% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM.test.R
\name{VECM.logl}
\alias{VECM.logl}
\alias{VECM.break.logl}
\title{Quasi log likelihood for VECM without and with breaks}
\usage{
VECM.logl(y, p)

VECM.break.logl(y, p, breaks.list)
}
\arguments{
\item{y}{A matrix of \eqn{n} VAR variables.}

\item{p}{Number of lags.}

\item{breaks.list}{Vector of possible break dates.}
}
\value{
The vector or matrix with the values of quasi log likelihood
for all possible values of cointegration rank and different break dates,
if the function with breaks is called.
}
\description{
Quasi log likelihood for VECM without and with breaks
}
\keyword{internal}
