% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECM.test.R
\name{VECM.test}
\alias{VECM.test}
\title{Test of the co-integration rank with a possible break in trend}
\usage{
VECM.test(y, r, max.lag, trim = 0.15)
}
\arguments{
\item{y}{The matrix of \eqn{n} VAR variables.}

\item{r}{The co-integration rank tested against the alternative of \eqn{n}.}

\item{max.lag}{The maximum number of lags.}

\item{trim}{The trimming parameter to determine the lower and upper bounds.}
}
\value{
A list of:
\itemize{
\item the indicator of the rejection of null.
\item the estimated break point.
\item the estimated lag number.
}
}
\description{
This procedure is aimed on the problem of testing for the co-integration rank
of a vector autoregressive process in the case where a trend break may
potentially be present in the data.

The test is based on estimating the quasi log likelihood for two situations,
with break, and without it. The one with the smallest value is considered to
be the result.
}
\references{
Harris, David, Stephen J. Leybourne, and A. M. Robert Taylor.
“Tests of the Co-Integration Rank in VAR Models in the Presence
of a Possible Break in Trend at an Unknown Point.”
Journal of Econometrics, Innovations in Multiple Time Series Analysis,
192, no. 2 (June 1, 2016): 451–67.
https://doi.org/10.1016/j.jeconom.2016.02.010.
}
