% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.GSADF.test.R
\name{weighted.GSADF.test}
\alias{weighted.GSADF.test}
\title{Weighted generalized supremum ADF test.}
\usage{
weighted.GSADF.test(
  y,
  trim = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  alpha = 0.05,
  iter = 4 * 200,
  urs = TRUE,
  seed = round(10^4 * sd(y))
)
}
\arguments{
\item{y}{The input time series of interest.}

\item{trim}{Trimming parameter to determine the lower and upper bounds.}

\item{const}{Whether the constant needs to be included.}

\item{alpha}{The significance level of interest.}

\item{iter}{The number of iterations.}

\item{urs}{Use `union of rejections` strategy.}

\item{seed}{The seed parameter for the random number generator.}
}
\description{
Weighted generalized supremum ADF test.
}
\references{
Harvey, David I., Stephen J. Leybourne, and Yang Zu.
“Testing Explosive Bubbles with Time-Varying Volatility.”
Econometric Reviews 38, no. 10 (November 26, 2019): 1131–51.
https://doi.org/10.1080/07474938.2018.1536099.

Kurozumi, Eiji, Anton Skrobotov, and Alexey Tsarev.
“Time-Transformed Test for the Explosive Bubbles under
Non-Stationary Volatility.”
arXiv, November 15, 2021. http://arxiv.org/abs/2012.13937.
}
