% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.SADF.test.R
\name{weighted.SADF.test}
\alias{weighted.SADF.test}
\title{Weighted SADF test (HLZ, 2018).}
\usage{
weighted.SADF.test(
  y,
  r0 = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  alpha = 0.05,
  iter = 4 * 200,
  urs = TRUE,
  seed = round(10^4 * sd(y))
)
}
\description{
Weighted SADF test (HLZ, 2018).
}
