% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ACF}
\alias{ACF}
\title{Calculating ACF values}
\usage{
ACF(y)
}
\arguments{
\item{y}{An input time series of interest.}
}
\value{
The vector of ACF values of orders from 0 to \eqn{N - 1}.
Due to the R's way of indexing the 1-st element is the autocorrelation of
order 0, and the N-th value is the autcorrelation of order \eqn{N - 1}.
}
\description{
A simple auxiliary function providing the estimates of autocorrelations of
orders from \eqn{0} to \eqn{N - 1}.
}
\keyword{internal}
