% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.test.R
\name{ADF.test}
\alias{ADF.test}
\title{ADF.test - a simple implementation of ADF test}
\usage{
ADF.test(
  y,
  const = TRUE,
  trend = FALSE,
  max.lag = 0,
  criterion = NULL,
  modified.criterion = FALSE,
  rescale.criterion = FALSE
)
}
\arguments{
\item{y}{The input time series of interest.}

\item{const, trend}{Whether a constand and trend are to be included.}

\item{max.lag}{Maximum lag number}

\item{criterion}{A criterion used to select number of lags.
If lag selection is not needed keep this NULL.}

\item{modified.criterion}{Whether the unit-root test modificaton is needed.}

\item{rescale.criterion}{Whether the rescaling informational criterion
is needed. Designed to cope with heteroscedasticity in residuals.}
}
\value{
List containing:
\itemize{
\item y
\item const
\item trend
\item residuals
\item coefficient estimates
\item t-statistic value
\item critical value
\item Number of lags
\item indicator of stationarity}
}
\description{
A function for ADF test with the ability to select the number of lags.
Lags are selected by informational criterions which can be modified as in
Ng and Perron (2001) and Cavaliere et al. (2015).
}
\details{
Due to the Frisch-Waugh-Lovell theorem we first detrend `y` and then apply
the test to the detrended series.
}
\references{
Cavaliere, Giuseppe, Peter C. B. Phillips, Stephan Smeekes,
and A. M. Robert Taylor. “Lag Length Selection for Unit Root Tests
in the Presence of Nonstationary Volatility.”
Econometric Reviews 34, no. 4 (April 21, 2015): 512–36.
https://doi.org/10.1080/07474938.2013.808065.

Ng, Serena, and Pierre Perron. “Lag Length Selection and the Construction of
Unit Root Tests with Good Size and Power.”
Econometrica 69, no. 6 (2001): 1519–54.
https://doi.org/10.1111/1468-0262.00256.
}
