% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.test.S.R
\name{ADF.test.S}
\alias{ADF.test.S}
\title{Detrending bootstrap test by Smeekes (2013)}
\usage{
ADF.test.S(
  y,
  const = TRUE,
  trend = FALSE,
  c = 0,
  gamma = 0,
  trim = 0.15,
  max.lag = 0,
  criterion = NULL,
  modified.criterion = FALSE,
  iter = 999
)
}
\arguments{
\item{y}{A time series of interest.}

\item{const, trend}{Whether the constant and trend are to be included.}

\item{c}{A filtration parameter used to construct an autocorrelation
coefficient.}

\item{gamma}{Detrending type selection parameter. If 0 the OLS detrending
is applied, if 1 the GLS detrending is applied, otherwise the autocorrelation
coefficient is calculated as \eqn{1 + c^{\gamma} T^{-\gamma}}.}

\item{trim}{A trimming parameter.}

\item{max.lag}{The maximum lag for inner ADF testing.}

\item{criterion}{A criterion used to select number of lags.
If lag selection is not needed keep this NULL.}

\item{modified.criterion}{Whether the unit-root test modificaton is needed.}

\item{iter}{The number of bootstrap iterations.}
}
\description{
This bootstrap test is based on the recursive detrending procedure of Taylor
(2002). The main idea is to apply the standard ADF test to the series
with nuissanse parameters eliminated.
}
\details{
Critical values are calculated via a bootstrapping using MacKinnon-like
regressions. For each number of observations and each number of variables
obtained were 1999 values of test statistics. After that 1st, 2.5-th, 5-th,
10-th, and 97.5-th percentiles were calculated and saved along with the
corresponding number of observations. This step was repeated 5 times to cope
with possible biases. After that MacKinnon-like regressions were estimated.
}
\references{
Taylor, A. M. Robert.
“Regression-Based Unit Root Tests With Recursive Mean Adjustment for
Seasonal and Nonseasonal Time Series.”
Journal of Business & Economic Statistics 20, no. 2 (April 2002): 269–81.
https://doi.org/10.1198/073500102317352001.

MacKinnon, James G.
“Critical Values for Cointegration Tests.”
Working Paper. Economics Department, Queen’s University, January 2010.
https://ideas.repec.org/p/qed/wpaper/1227.html.

Smeekes, Stephan.
“Detrending Bootstrap Unit Root Tests.”
Econometric Reviews 32, no. 8 (July 2013): 869–91.
https://doi.org/10.1080/07474938.2012.690693.

Elliott, Graham, Thomas J. Rothenberg, and James H. Stock.
“Efficient Tests for an Autoregressive Unit Root.”
Econometrica 64, no. 4 (1996): 813–36.
https://doi.org/10.2307/2171846.
}
