% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.test.S.R
\name{ADF.test.S}
\alias{ADF.test.S}
\title{Detrending bootstrap test by Smeekes (2013)}
\usage{
ADF.test.S(
  y,
  const = TRUE,
  trend = FALSE,
  c = 0,
  gamma = 0,
  trim = 0.15,
  max.lag = 0,
  criterion = NULL,
  modified.criterion = FALSE,
  iter = 999
)
}
\arguments{
\item{y}{The series of interest.}

\item{const}{Whether the constant is to included.}

\item{trend}{Whether the trend is to be included.}

\item{c}{A filtration parameter.}

\item{gamma}{A detrending type selection parameter. If 0 the OLS detrending
is applied, if 1 the GLS detrending is applied, otherwise the autocorrelation
coefficient is calculated as \eqn{1 + с^{\gamma} T^{\gamma}}.}

\item{trim}{The trimming parameter.}

\item{max.lag}{The maximum lag for inner ADF testing.}

\item{criterion}{A criterion used to select number of lags.
If lag selection is not needed keep this NULL.}

\item{modified.criterion}{Whether the unit-root test modificaton is needed.}

\item{iter}{The number of bootstrap steps.}
}
\description{
Detrending bootstrap test by Smeekes (2013)
}
\references{
Smeekes, Stephan.
“Detrending Bootstrap Unit Root Tests.”
Econometric Reviews 32, no. 8 (July 2013): 869–91.
https://doi.org/10.1080/07474938.2012.690693.
}
