% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrews_kim_test.R
\name{andrews_kim_test}
\alias{andrews_kim_test}
\title{Andrews-Kim (2006) test}
\usage{
andrews_kim_test(eq, m, dataset)
}
\arguments{
\item{eq}{Base model formula}

\item{m}{Post-break period length}

\item{dataset}{Source of the data}
}
\value{
The list, containing
\itemize{
\item **m**
\item estimated values of P- and R-tests.
\item sequences of auxiliary statistics \eqn{P_j} and \eqn{R_j}.
\item the corresponding p-values.
}
}
\description{
Test for structural break at the sample end
}
\details{
See Andrews and Kim (2006) for the detailed description.
}
