% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS-GLS.R
\name{AR}
\alias{AR}
\title{Custom AR with extra information}
\usage{
AR(y, x, max.lag, criterion = "aic")
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Exogenous explanatory variables.}

\item{max.lag}{The maximum number of lags.}

\item{criterion}{A criterion for lag number estimation.}
}
\value{
A list of:
\itemize{
\item \code{beta}: estimates of coefficients,
\item \code{residuals}: estimated residuals,
\item \code{predict}: forecasted values,
\item \code{t.beta}: \eqn{t}-statistics for \code{beta},
\item \code{lag}: estimated number of lags.
}
}
\description{
Custom AR with extra information
}
\details{
The function is not intended to be used directly so it's not exported.
}
