% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coint.test.GH.R
\name{coint.test.GH}
\alias{coint.test.GH}
\title{Gregory-Hansen test for the absense of cointegration}
\usage{
coint.test.GH(..., shift = "level", trim = 0.15, max.lag = 10)
}
\arguments{
\item{...}{Variables of interest.}

\item{shift}{Expected break type.}

\item{trim}{The trimming parameter to calculate break moment bounds.}

\item{max.lag}{The maximum number of lags for the internal ADF testing.}
}
\description{
Gregory-Hansen test for the absense of cointegration under the assumption
of a structural break at the unknown moment of time.
}
\references{
Gregory, Allan W., and Bruce E. Hansen.
“Residual-Based Tests for Cointegration in Models with Regime Shifts.”
Journal of Econometrics 70, no. 1 (January 1, 1996): 99–126.
https://doi.org/10.1016/0304-4076(69)41685-7.
}
