% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determi_kpss_2p.R
\name{determi_kpss_2p}
\alias{determi_kpss_2p}
\title{Procedure to compute deterministic terms
for KPSS with 2 structural breaks.}
\usage{
determi_kpss_2p(model, N, break.point)
}
\arguments{
\item{model}{\describe{
\item{1}{for the AA (without trend) model.}
\item{2}{for the AA (with trend) model.}
\item{3}{for the BB model.}
\item{4}{for the CC model.}
\item{5}{for the AC-CA model.}
}}

\item{N}{Number of observations.}

\item{break.point}{Positions for the first and second structural breaks
(respective to the origin which is 1).}
}
\value{
Matrix of deterministic terms.
}
\description{
Procedure to compute deterministic terms
for KPSS with 2 structural breaks.
}
