% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinants.kpss.N.break.R
\name{determinants.kpss.N.breaks}
\alias{determinants.kpss.N.breaks}
\title{Deterministic terms for multi-break KPSS}
\usage{
determinants.kpss.N.breaks(model, N, break.point, const = FALSE, trend = FALSE)
}
\arguments{
\item{model}{A scalar or vector of
\describe{
\item{1}{for the break in const.}
\item{2}{for the break in trend.}
\item{3}{for the break in const and trend.}
}}

\item{N}{Number of observations.}

\item{break.point}{Array of structural breaks.}

\item{const}{Include constant if **TRUE**.}

\item{trend}{Include trend if **TRUE**.}
}
\value{
Matrix of deterministic terms.
}
\description{
Procedure to compute deterministic terms
for KPSS with \eqn{m} structural breaks.
}
\details{
**model** should be either a scalar or a vector of the same size as the
**break.point**. If scalar **model** will be repeated till the length of
**break.point** is achieved.
}
