% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDF.CHLT.R
\name{GLS.bt}
\alias{GLS.bt}
\title{GLS fitering with a break}
\usage{
GLS.bt(y, lambda, c)
}
\arguments{
\item{y}{Series of interest.}

\item{lambda}{Break relative position.}

\item{c}{Filtering parameter.}
}
\description{
GLS fitering with a break
}
\keyword{internal}
