% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.Estimates.R
\name{GLS}
\alias{GLS}
\title{Custom GLS with extra information}
\usage{
GLS(y, z, c)
}
\arguments{
\item{y}{Dependent variable.}

\item{z}{Explanatory variables.}

\item{c}{Coefficient for \eqn{\rho} calculation.}
}
\value{
The list of betas, residuals, forecasted values and t-values.
}
\description{
Getting GLS estimates of betas, residuals, forecasted values and t-values.
}
\details{
The function is not intended to be used directly so it's not exported.
}
