% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSADF.bootstrap.test.R
\name{GSADF.bootstrap.test}
\alias{GSADF.bootstrap.test}
\title{Generalized supremum ADF test with wild bootstrap.}
\usage{
GSADF.bootstrap.test(
  y,
  trim = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  alpha = 0.05,
  iter = 4 * 200,
  seed = round(10^4 * sd(y))
)
}
\arguments{
\item{y}{The input time series of interest.}

\item{trim}{Trimming parameter to determine the lower and upper bounds.}

\item{const}{Whether the constant needs to be included.}

\item{alpha}{The significance level of interest.}

\item{iter}{The number of iterations.}

\item{seed}{The seed parameter for the random number generator.}
}
\description{
Generalized supremum ADF test with wild bootstrap.
}
\references{
Kurozumi, Eiji, Anton Skrobotov, and Alexey Tsarev.
“Time-Transformed Test for Bubbles under Non-Stationary Volatility.”
Journal of Financial Econometrics, April 23, 2022.
https://doi.org/10.1093/jjfinec/nbac004.
}
