% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.criterion.R
\name{info.criterion}
\alias{info.criterion}
\title{Information criterions}
\usage{
info.criterion(resid, extra, modification = FALSE, ...)
}
\arguments{
\item{resid}{Input residuals.}

\item{extra}{Number of coefficients etc.}

\item{modification}{Whether the unit-root test modificaton is needed.}

\item{...}{Extra parameters needed for modification. Needed are
\itemize{
\item The coefficient \eqn{\alpha} of \eqn{y_{t-1}},
\item The vector of \eqn{y_{t-1}},
}
exactly in this order.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}
}
\value{
The value of the selected informational criterion.
}
\description{
Information criterions
}
\details{
Calculating the value of the following informational criterions:
\itemize{
\item Akaike,
\item Schwarz (Bayesian),
\item Hannan-Quinn,
\item Liu et al.}
}
