% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KP.R
\name{KP}
\alias{KP}
\title{Kejrival-Perron procedure of breaks number detection}
\usage{
KP(y, const = FALSE, breaks = 1, criterion = "aic", trim = 0.15)
}
\arguments{
\item{y}{An input series of interest.}

\item{const}{Whether the break in constant is allowed.}

\item{breaks}{Number of breaks.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}

\item{trim}{A trimming value for a possible break date bounds.}
}
\value{
The estimated optimal break point.
}
\description{
Kejrival-Perron procedure of breaks number detection
}
\references{
Kejriwal, Mohitosh, and Pierre Perron.
“A Sequential Procedure to Determine the Number of Breaks in Trend
with an Integrated or Stationary Noise Component:
Determination of Number of Breaks in Trend.”
Journal of Time Series Analysis 31, no. 5 (September 2010): 305–28.
https://doi.org/10.1111/j.1467-9892.2010.00666.x.
}
