% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss.1.break.R
\name{kpss.1.break}
\alias{kpss.1.break}
\title{KPSS-test with known structural break}
\usage{
kpss.1.break(y, x, model, break.point, weakly.exog = TRUE, ll.init)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable.}

\item{x}{(Txk)-matrix of explanatory stochastic regressors.}

\item{model}{\describe{
\item{1}{for model An.}
\item{2}{for model A.}
\item{3}{for model B.}
\item{4}{for model C.}
\item{5}{for model D.}
\item{6}{for model E.}
}}

\item{break.point}{Position of the break point.}

\item{weakly.exog}{Exogeneity of the stochastic regressors
\describe{
\item{TRUE}{if the regressors are weakly exogenous,}
\item{FALSE}{if the regressors are not weakly exogenous
(DOLS is used in this case).}
}}

\item{ll.init}{Scalar, defines the initial number of leads and lags for DOLS.}
}
\value{
\describe{
\item{beta}{DOLS estimates of the coefficients regressors.}
\item{tests}{SC test (coinkpss-test).}
\item{resid}{Residuals of the model.}
\item{t.beta}{Individual significance t-statistics.}
\item{break_point}{Break points.}
}
}
\description{
Computes the cointegration test with one known structural break.
}
\details{
The code provided is the original GAUSS code ported to R.

See Carrion-i-Silvestre and Sansó (2006) for further details.
}
