% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.2.breaks.R
\name{KPSS.2.breaks}
\alias{KPSS.2.breaks}
\title{KPSS-test with 2 known structural breaks}
\usage{
KPSS.2.breaks(y, model, break.point, max.lag, kernel)
}
\arguments{
\item{y}{An input (LHS) time series of interest.}

\item{model}{A scalar equal to
\itemize{
\item 1: for the AA (without trend) model,
\item 2: for the AA (with trend) model,
\item 3: for the BB model,
\item 4: for the CC model,
\item 5: for the AC-CA model.
}}

\item{break.point}{Positions for the first and second structural breaks
(respective to the origin which is 1).}

\item{max.lag}{scalar, with the maximum order of the parametric correction.
The final order of the parametric correction is selected using the
BIC information criterion.}

\item{kernel}{Kernel for calculating long-run variance
\itemize{
\item \code{bartlett}: for Bartlett kernel,
\item \code{quadratic}: for Quadratic Spectral kernel,
\item \code{NULL} for the Kurozumi's proposal, using Bartlett kernel.
}}
}
\value{
A list of:
\itemize{
\item \code{beta}: DOLS estimates of the coefficients,
\item \code{tests}: SC test (coinKPSS-test),
\item \code{resid}: Residuals of the model,
\item \code{t.beta}: \eqn{t}-statistics for \code{beta},
\item \code{break_point}: Break points.
}
}
\description{
Procedure to compute the KPSS test with two structural breaks
}
\details{
The break points are known

The code provided is the original GAUSS code ported to R.

See Carrion-i-Silvestre and Sansó (2007) for further details.
}
\references{
Carrion-i-Silvestre, Josep Lluís, and Andreu Sansó.
“The KPSS Test with Two Structural Breaks.”
Spanish Economic Review 9, no. 2 (May 16, 2007): 105–27.
https://doi.org/10.1007/s10108-006-9017-8.
}
