% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.2.breaks.R
\name{KPSS.2.breaks.unknown}
\alias{KPSS.2.breaks.unknown}
\title{KPSS-test with 2 unknown structural breaks}
\usage{
KPSS.2.breaks.unknown(y, model, max.lag = 0, kernel = "bartlett")
}
\arguments{
\item{y}{A time series of interest.}

\item{model}{A scalar equal to
\itemize{
\item 1: for the AA (without trend) model,
\item 2: for the AA (with trend) model,
\item 3: for the BB model,
\item 4: for the CC model,
\item 5: for the AC-CA model.
}}

\item{max.lag}{A scalar, with the maximum order of the parametric correction.
The final order of the parametric correction is selected using
the BIC information criterion.}

\item{kernel}{Kernel for calculating long-run variance
\itemize{
\item \code{bartlett}: for Bartlett kernel,
\item \code{quadratic}: for Quadratic Spectral kernel,
\item \code{NULL} for the Kurozumi's proposal, using Bartlett kernel.
}}
}
\value{
Value of test statistic.
}
\description{
Procedure to compute the KPSS test with two structural breaks
}
\details{
The break points are known

The code provided is the original GAUSS code ported to R.

See Carrion-i-Silvestre and Sansó (2007) for further details.
}
\references{
Carrion-i-Silvestre, Josep Lluís, and Andreu Sansó.
“Testing the Null of Cointegration with Structural Breaks.”
Oxford Bulletin of Economics and Statistics 68, no. 5 (October 2006): 623–46.
https://doi.org/10.1111/j.1468-0084.2006.00180.x.
}
