% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.N.breaks.bootstrap.R
\name{KPSS.N.breaks.bootstrap}
\alias{KPSS.N.breaks.bootstrap}
\title{KPSS-test with multiple unknown structural breaks}
\usage{
KPSS.N.breaks.bootstrap(
  y,
  x,
  model,
  break.point,
  const = FALSE,
  trend = FALSE,
  weakly.exog = TRUE,
  lags.init,
  leads.init,
  max.lag,
  kernel,
  iter = 9999,
  bootstrap = "sample",
  criterion = "bic"
)
}
\arguments{
\item{y}{An input (LHS) time series of interest.}

\item{x}{A matrix of (RHS) explanatory stochastic regressors.}

\item{model}{A scalar or vector of
\describe{
\item{1}{for the break in const.}
\item{2}{for the break in trend.}
\item{3}{for the break in const and trend.}
}}

\item{break.point}{Array of structural breaks.}

\item{const}{Include constant if **TRUE**.}

\item{trend}{Include trend if **TRUE**.}

\item{weakly.exog}{Exogeneity of the stochastic regressors
\describe{
\item{TRUE}{if the regressors are weakly exogenous,}
\item{FALSE}{if the regressors are not weakly exogenous
(DOLS is used in this case).}
}}

\item{lags.init, leads.init}{Scalars defininig the initial number of lags and
leads for DOLS.}

\item{max.lag}{scalar, with the maximum order of the parametric correction.
The final order of the parametric correction is selected using the BIC
information criterion.}

\item{kernel}{Kernel for calculating long-run variance
\describe{
\item{bartlett}{for Bartlett kernel.}
\item{quadratic}{for Quadratic Spectral kernel.}
\item{NULL}{for the Kurozumi's proposal, using Bartlett kernel.}
}}

\item{iter}{Number of bootstrap iterations.}

\item{bootstrap}{Type of bootstrapping:
\describe{
\item{sample}{sampling from residuals with replacement.}
\item{Cavaliere-Taylor}{multiplying residuals by \eqn{N(0, 1)}-distributed
variable.}
\item{Rademacher}{multiplying residuals by Rademacher-distributed variable.}
}}

\item{criterion}{Information criterion for DOLS lags and leads selection:
aic, bic or lwz.}
}
\value{
List of 3 elements:
\describe{
\item{test}{The value of KPSS test statistic.}
\item{p.value}{The estimates p-value.}
\item{bootstrapped}{Bootstrapped auxiliary statistics.}}
}
\description{
Procedure to compute the KPSS test with multiple unknown structural breaks
}
