% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.N.breaks.R
\name{KPSS.N.breaks}
\alias{KPSS.N.breaks}
\title{KPSS-test with multiple known structural breaks}
\usage{
KPSS.N.breaks(
  y,
  x,
  model,
  break.point,
  const = FALSE,
  trend = FALSE,
  weakly.exog = TRUE,
  lags.init,
  leads.init,
  max.lag,
  kernel,
  criterion = "bic"
)
}
\arguments{
\item{y}{A time series of interest.}

\item{x}{A matrix of explanatory stochastic regressors.}

\item{model}{A scalar or vector of
\itemize{
\item 1: for the break in const,
\item 2: for the break in trend,
\item 3: for the break in const and trend.
}}

\item{break.point}{Array of structural breaks.}

\item{const, trend}{Whether a constant or trend should be included.}

\item{weakly.exog}{Boolean where we specify
whether the stochastic regressors are exogenous or not
\itemize{
\item \code{TRUE}: if the regressors are weakly exogenous,
\item \code{FALSE}: if the regressors are not weakly exogenous
(DOLS is used in this case).
}}

\item{lags.init, leads.init}{Scalars defininig the initial number of lags and
leads for DOLS.}

\item{max.lag}{scalar, with the maximum order of the parametric correction.
The final order of the parametric correction is selected
using the BIC information criterion.}

\item{kernel}{Kernel for calculating long-run variance
\itemize{
\item \code{bartlett}: for Bartlett kernel,
\item \code{quadratic}: for Quadratic Spectral kernel,
\item \code{NULL} for the Kurozumi's proposal, using Bartlett kernel.
}}

\item{criterion}{Information criterion for DOLS lags and leads selection:
aic, bic, hq, or lwz.}
}
\value{
A list of
\itemize{
\item \code{beta}: DOLS estimates of the coefficients,
\item \code{tests}: SC test (coinKPSS-test),
\item \code{resid}: Residuals of the model,
\item \code{t.beta}: \eqn{t}-statistics for \code{beta},
\item \code{DOLS.lags}: The estimated number of lags and leads in DOLS,
\item \code{break_point}: Break points.
}
}
\description{
Procedure to compute the KPSS test with multiple known structural breaks
}
\references{
Carrion-i-Silvestre, Josep Lluís, and Andreu Sansó.
“Testing the Null of Cointegration with Structural Breaks.”
Oxford Bulletin of Economics and Statistics 68, no. 5 (October 2006): 623–46.
https://doi.org/10.1111/j.1468-0084.2006.00180.x.

Carrion-i-Silvestre, Josep Lluís, and Andreu Sansó.
“The KPSS Test with Two Structural Breaks.”
Spanish Economic Review 9, no. 2 (May 16, 2007): 105–27.
https://doi.org/10.1007/s10108-006-9017-8.
}
