% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.R
\name{KPSS}
\alias{KPSS}
\title{Auxiliary function returning KPSS statistic value.}
\usage{
KPSS(resids, variance)
}
\arguments{
\item{resids}{The series of residuals.}

\item{variance}{The value of the long-run variance.}
}
\description{
Auxiliary function returning KPSS statistic value.
}
\details{
The function is not intended to be used directly so it's not exported.
}
