% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPSS.R
\name{KPSS}
\alias{KPSS}
\title{Auxiliary function returning KPSS statistic value.}
\usage{
KPSS(resid, variance)
}
\arguments{
\item{resid}{The series of residuals.}

\item{variance}{The value of the long-run variance.}
}
\description{
Auxiliary function returning KPSS statistic value.
}
