% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss_unknown_1p.R
\name{kpss_unknown_1p}
\alias{kpss_unknown_1p}
\title{KPSS-test of cointegration}
\usage{
kpss_unknown_1p(y, x, model, weakly.exog, k)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable}

\item{x}{(Txk)-matrix of explanatory stochastic regressors}

\item{model}{Scalar, denotes the deterministic component:
\describe{
\item{1}{for model An.}
\item{2}{for model A.}
\item{3}{for model B.}
\item{4}{for model C.}
\item{5}{for model D.}
\item{6}{for model E.}
}}

\item{weakly.exog}{Boolean where we specify whether the stochastic regressors are exogenous or not
\describe{
\item{TRUE}{if the regressors are weakly exogenous,}
\item{FALSE}{if the regressors are not weakly exogenous (DOLS is used in this case).}
}}

\item{k}{Scalar, defines the initial number of leads and lags for DOLS.}
}
\value{
vec_out (2x2)-matrix, where the first rows gives the value of the min(SC) test and the estimated break point; the second row gives the value of the SC statistic, where the break point is estimated as min(SSR).
}
\description{
Procedure for testing the null of cointegration in the possible presence of structural breaks
}
\details{
Computes the cointegration test with one unknown structural break where the break point is estimated either minimizing the value of the statistic or the sum of the squared residuals. The estimation of the cointegrating relationship bases on DOLS.

The code provided is the original GAUSS code ported to R.

See Carrion-i-Silvestre and Sansó (2006) for further details.
}
