% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.var.R
\name{lr.var.bartlett}
\alias{lr.var.bartlett}
\title{Calculating long-run variance with Bartlett kernel}
\usage{
lr.var.bartlett(y, l = NULL)
}
\arguments{
\item{y}{A series of interest.}

\item{l}{Number of lags/leads. If \code{NULL} then it's estimated.}
}
\value{
Long-run variance.
}
\description{
Calculating long-run variance with Bartlett kernel
}
\details{
The function is not intended to be used directly so it's not exported.
}
