% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.var.R
\name{lr.var.quadratic}
\alias{lr.var.quadratic}
\title{Calculating long-run variance with quadratic kernel}
\usage{
lr.var.quadratic(y)
}
\arguments{
\item{y}{A series of interest.}
}
\description{
Calculating long-run variance with quadratic kernel
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Andrews, Donald W. K.
“Heteroskedasticity and Autocorrelation Consistent
Covariance Matrix Estimation.”
Econometrica 59, no. 3 (1991): 817–58.
https://doi.org/10.2307/2938229.
}
