% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.var.R
\name{lr.var.SPC}
\alias{lr.var.SPC}
\title{Calculating long-run variance}
\usage{
lr.var.SPC(y, max.lag = 0, kernel = "bartlett", criterion = "bic")
}
\arguments{
\item{y}{(Tx1) vector or residuals.}

\item{max.lag}{Maximum number of lags.
The exact number is selected by information criterions.}

\item{kernel}{Kernel for calculating long-run variance
\describe{
\item{bartlett}{for Bartlett kernel.}
\item{quadratic}{for Quadratic Spectral kernel.}
\item{NULL}{for the Kurozumi's proposal, using Bartlett kernel.}
}}

\item{criterion}{The information crietreion: bic, aic or lwz.}
}
\value{
Long-run variance.
}
\description{
Procedure ALRVR to estimate the long-run variance as
in Sul, Phillips and Choi (2003).
}
\details{
Used are Quadratic Spectral and Bartlett kernels.
}
\references{
Sul, Donggyu, Peter C. B. Phillips, and Chi-Young Choi.
“Prewhitening Bias in HAC Estimation.”
Oxford Bulletin of Economics and Statistics 67, no. 4 (August 2005): 517–46.
https://doi.org/10.1111/j.1468-0084.2005.00130.x.
}
