% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDF.multiple.R
\name{MDF.multiple}
\alias{MDF.multiple}
\title{MDF procedure for multiple unknown breaks.}
\usage{
MDF.multiple(
  y,
  const = FALSE,
  breaks = 1,
  breaks.star = 1,
  trim = 0.15,
  ZA = FALSE
)
}
\arguments{
\item{y}{The series of interest.}

\item{const}{Whether the constant term should be included.}

\item{breaks}{Number of breaks.}

\item{breaks.star}{Number of breaks got from the Kejrival-Perron procedure.}

\item{trim}{A trimming value for a possible break date bounds.}

\item{ZA}{Whether ZA variant should be used.}
}
\description{
MDF procedure for multiple unknown breaks.
}
