% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.Estimates.R
\name{NW.kernel}
\alias{NW.kernel}
\title{Returning a kernel value for a specific point}
\usage{
NW.kernel(i, x, h, kernel = "unif")
}
\arguments{
\item{i}{Index of the base point.}

\item{x}{Series for kernel calculations.}

\item{h}{Kernel bandwidth.}

\item{kernel}{Kernel type:
\itemize{
\item \code{unif}: \eqn{K(x) = \left\{\begin{array}{ll}
1 & \frac{|x - x_i|}{h} \leq 1 \\
0 & \textrm{otherwize}
\end{array}\right.}
\item \code{gauss}: \eqn{\Phi(\frac{x - x_i}{h})}
}}
}
\description{
Returning a kernel value for a specific point
}
\keyword{internal}
