% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW.estimation.R
\name{NW.kernel}
\alias{NW.kernel}
\title{Nadaraya–Watson kernel series}
\usage{
NW.kernel(i, x, h, kernel = "unif")
}
\arguments{
\item{i}{Current index.}

\item{x}{Kernel variable.}

\item{h}{Bandwidth.}

\item{kernel}{Needed kernel, currently only \code{unif} and \code{gauss}.}
}
\value{
A series of kernel function values.
}
\description{
Nadaraya–Watson kernel series
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Harvey, David I., S. Leybourne, Stephen J., and Yang Zu.
“Nonparametric Estimation of the Variance Function
in an Explosive Autoregression Model.”
School of Economics. University of Nottingham, 2022.
}
