% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.Estimates.R
\name{NW.loocv}
\alias{NW.loocv}
\title{LOO-CV for h in Nadaraya–Watson kernel regression.}
\usage{
NW.loocv(y, x, kernel = "unif")
}
\arguments{
\item{y}{LHS dependent variable.}

\item{x}{RHS explanation variable.}

\item{kernel}{Needed kernel, currently only \code{unif} and \code{gauss}.}
}
\value{
A list of arguments as well as the estimated bandwidth \code{h}.
}
\description{
LOO-CV for h in Nadaraya–Watson kernel regression.
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Harvey, David I., S. Leybourne, Stephen J., and Yang Zu.
“Nonparametric Estimation of the Variance Function
in an Explosive Autoregression Model.”
School of Economics. University of Nottingham, 2022.
}
