% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW.estimation.R
\name{NW.loocv}
\alias{NW.loocv}
\title{NW.loocv - LOO-CV for h in Nadaraya–Watson kernel regression.}
\usage{
NW.loocv(y, x, kernel = "unif")
}
\arguments{
\item{y}{LHS dependent variable.}

\item{x}{RHS explanation variable.}

\item{kernel}{Needed kernel, currently only `unif` and `gauss`.}
}
\value{
A list of arguments as well as the estimated bandwidth `h`.
}
\description{
NW.loocv - LOO-CV for h in Nadaraya–Watson kernel regression.
}
\references{
Harvey, David I., S. Leybourne, Stephen J., and Yang Zu.
“Nonparametric Estimation of the Variance Function
in an Explosive Autoregression Model.”
School of Economics. University of Nottingham, 2022.
}
