% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW.estimation.R
\name{NW.volatility}
\alias{NW.volatility}
\title{NW.volatility - Nadaraya–Watson kernel volatility estimation}
\usage{
NW.volatility(e, h, kernel = "unif")
}
\arguments{
\item{e}{The series of interest.}

\item{h}{Bandwidth.}

\item{kernel}{Needed kernel, currently only \code{unif} and \code{gauss}.}
}
\value{
A list of arguments as well as the estimated omega and s.e.
}
\description{
NW.volatility - Nadaraya–Watson kernel volatility estimation
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Cavaliere, Giuseppe, Peter C. B. Phillips, Stephan Smeekes,
and A. M. Robert Taylor.
“Lag Length Selection for Unit Root Tests in the Presence
of Nonstationary Volatility.”
Econometric Reviews 34, no. 4 (April 21, 2015): 512–36.
https://doi.org/10.1080/07474938.2013.808065.

Harvey, David I., S. Leybourne, Stephen J., and Yang Zu.
“Nonparametric Estimation of the Variance Function
in an Explosive Autoregression Model.”
School of Economics. University of Nottingham, 2022.
}
