% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS-GLS.R
\name{OLS}
\alias{OLS}
\title{Custom OLS with extra information}
\usage{
OLS(y, x)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Explanatory variables.}
}
\value{
The list of betas, residuals, forecasted values and t-values.
}
\description{
Getting OLS estimates of betas, residuals, forecasted values and t-values.
}
\details{
The function is not intended to be used directly so it's not exported.
}
