% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsqr.R
\name{olsqr}
\alias{olsqr}
\title{Getting OLS estimates of betas, residuals, forecasted values and t-values.}
\usage{
olsqr(y, x)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Explanatory variables.}
}
\value{
The list of betas, residuals, forecasted values and t-values.
}
\description{
Getting OLS estimates of betas, residuals, forecasted values and t-values.
}
