% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.values.R
\name{p.values.SADF}
\alias{p.values.SADF}
\title{Critical values for SADF-type tests}
\usage{
p.values.SADF(statistic, N.obs, cr.values)
}
\arguments{
\item{statistic}{The statistic value.}

\item{N.obs}{The number of observations.}

\item{cr.values}{The set of precalculated tables.}
}
\description{
Interpolating p-value for intermediate observation numbers for SADF-type
tests.
}
\details{
The function is not intended to be used directly so it's not exported.
}
