% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.results.R
\name{print.sadf}
\alias{print.sadf}
\alias{print.mdfHLT}
\alias{print.mdfHLTN}
\alias{print.mdfCHLT}
\alias{print.cointGH}
\alias{print.robustUR}
\alias{print.robustURN}
\title{Custom functions for printing results in a nice way.}
\usage{
\method{print}{sadf}(x, ...)

\method{print}{mdfHLT}(x, ...)

\method{print}{mdfHLTN}(x, ...)

\method{print}{mdfCHLT}(x, ...)

\method{print}{cointGH}(x, ...)

\method{print}{robustUR}(x, ...)

\method{print}{robustURN}(x, ...)
}
\arguments{
\item{x}{An jbject containing results.}

\item{...}{Any additional arguments for \link{print} function.}
}
\description{
Custom functions for printing results in a nice way.
}
\keyword{internal}
