% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PY.sequential.R
\name{PY.sequential}
\alias{PY.sequential}
\title{Sequential Perron-Yabu (2009) statistic for breaks at unknown date.}
\usage{
PY.sequential(
  y,
  const = FALSE,
  breaks = 1,
  criterion = "aic",
  trim = 0.15,
  max.lag = 1
)
}
\arguments{
\item{y}{The input series of interest.}

\item{const}{Allowing the break in constant.}

\item{breaks}{Number of breaks.}

\item{criterion}{Needed information criterion: aic, bic, hq or lwz.}

\item{trim}{A trimming value for a possible break date bounds.}

\item{max.lag}{The maximum possible lag in the model.}
}
\description{
Sequential Perron-Yabu (2009) statistic for breaks at unknown date.
}
\references{
Kejriwal, Mohitosh, and Pierre Perron.
“A Sequential Procedure to Determine the Number of Breaks in Trend
with an Integrated or Stationary Noise Component:
Determination of Number of Breaks in Trend.”
Journal of Time Series Analysis 31, no. 5 (September 2010): 305–28.
https://doi.org/10.1111/j.1467-9892.2010.00666.x.
}
