% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindex.R
\name{reindex}
\alias{reindex}
\title{A function that makes reindexing}
\usage{
reindex(u)
}
\arguments{
\item{u}{The residuals series for reindexing.}
}
\description{
The function is aimed to calculate the sequence of indices providing a new
"time transformed" time series as in Cavaliere and Taylor (2008).
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Cavaliere, Giuseppe, and A. M. Robert Taylor.
“Time-Transformed Unit Root Tests for Models with Non-Stationary Volatility.”
Journal of Time Series Analysis 29, no. 2 (March 2008): 300–330.
https://doi.org/10.1111/j.1467-9892.2007.00557.x.

Kurozumi, Eiji, Anton Skrobotov, and Alexey Tsarev.
“Time-Transformed Test for Bubbles under Non-Stationary Volatility.”
Journal of Financial Econometrics, April 23, 2022.
https://doi.org/10.1093/jjfinec/nbac004.
}
