% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.tests.R
\name{robust.tests.single}
\alias{robust.tests.single}
\title{A wrapping function around \link{MDF.single}.}
\usage{
robust.tests.single(
  y,
  const = FALSE,
  trend = FALSE,
  season = FALSE,
  trim = 0.15
)
}
\arguments{
\item{y}{A series of interest.}

\item{const}{Whether the constant term should be included.}

\item{trend}{Whether the trend term should be included.}

\item{season}{Whether the seasonal adjustment is needed.}

\item{trim}{Trimming value for a possible break date bounds.}
}
\description{
A wrapping function around \link{MDF.single}.
}
