% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SADF.bootstrap.test.R
\name{SADF.bootstrap.test}
\alias{SADF.bootstrap.test}
\alias{GSADF.bootstrap.test}
\title{Supremum ADF tests with wild bootstrap.}
\usage{
SADF.bootstrap.test(
  y,
  trim = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  alpha = 0.05,
  iter = 999,
  seed = round(10^4 * sd(y))
)

GSADF.bootstrap.test(
  y,
  trim = 0.01 + 1.8/sqrt(length(y)),
  const = TRUE,
  alpha = 0.05,
  iter = 4 * 200,
  seed = round(10^4 * sd(y))
)
}
\arguments{
\item{y}{A time series of interest.}

\item{trim}{A trimming parameter to determine the lower and upper bounds for
a possible break point.}

\item{const}{Whether the constant needs to be included.}

\item{alpha}{The significance level of interest.}

\item{iter}{The number of iterations.}

\item{seed}{The seed parameter for the random number generator.}
}
\value{
An object of type \code{sadf}. It's a list of:
\itemize{
\item \code{y},
\item \code{trim},
\item \code{const},
\item \code{alpha},
\item \code{iter},
\item \code{seed},
\item vector of \eqn{t}-values,
\item the value of the corresponding test statistic,
\item series of bootstrapped test statistics,
\item bootstrapped critical values,
\item \eqn{p}-value.
}
}
\description{
\code{SADF.bootstrap.test} is a wild bootstrapping procedure for estimating
critical and \eqn{p}-values for \link{SADF.test}.

\code{GSADF.bootstrap.test} is the same procedure but for \link{GSADF.test}.
}
\details{
Refactored original code by Kurozumi et al.
}
\references{
Kurozumi, Eiji, Anton Skrobotov, and Alexey Tsarev.
“Time-Transformed Test for Bubbles under Non-Stationary Volatility.”
Journal of Financial Econometrics, April 23, 2022.
https://doi.org/10.1093/jjfinec/nbac004.
}
