% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinants.R
\name{seasonal.dummies}
\alias{seasonal.dummies}
\title{Generating seasomal monthly dummy variables.}
\usage{
seasonal.dummies(N)
}
\arguments{
\item{N}{number of observations.}
}
\value{
The matrix of values od seasonal dummies.
}
\description{
Generating seasomal monthly dummy variables.
}
