% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.GLS.R
\name{segments.GLS}
\alias{segments.GLS}
\title{Procedure to minimize the GLS-SSR for 1 break point}
\usage{
segments.GLS(
  y,
  const = FALSE,
  trend = FALSE,
  breaks = 1,
  first.break = NULL,
  last.break = NULL,
  trim = 0.15
)
}
\arguments{
\item{y}{Variable of interest.}

\item{const}{Whether there is a break in the constant.}

\item{trend}{Whether there is a break in the trend.}

\item{breaks}{Number of breaks.}

\item{first.break}{First possible break point.}

\item{last.break}{Last possible break point.}

\item{trim}{Trim value to calculate \code{first.break} and \code{last.break}
if not provided.}
}
\value{
The point of possible break.
}
\description{
Procedure to minimize the GLS-SSR for 1 break point
}
\details{
The function is not intended to be used directly so it's not exported.
}
\references{
Skrobotov, Anton.
“On Trend Breaks and Initial Condition in Unit Root Testing.”
Journal of Time Series Econometrics 10, no. 1 (2018): 1–15.
https://doi.org/10.1515/jtse-2016-0014.
}
