% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.OLS.R
\name{segments.OLS}
\alias{segments.OLS}
\title{Find \eqn{m + 1} optimal partitions}
\usage{
segments.OLS(y, x, m = 1, width = 2, SSR.data = NULL)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable.}

\item{x}{(Txk)-vector of the explanatory stochastic regressors.}

\item{m}{Number of breaks.}

\item{width}{Minimum spacing between the breaks.}

\item{SSR.data}{Optional matrix of recursive SSR's.}
}
\value{
List of 2 elements: optimal SSR and the vector of break points.
}
\description{
Find \eqn{m + 1} optimal partitions
}
\details{
Based on Bai and Perron (2003).
}
