% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.OLS.single.R
\name{segments.OLS.single}
\alias{segments.OLS.single}
\title{Procedure to minimize the SSR for 1 break point}
\usage{
segments.OLS.single(beg, end, first.break, last.break, len, SSR.data)
}
\arguments{
\item{beg}{Sample begin.}

\item{end}{Sample end.}

\item{first.break}{First possible break point.}

\item{last.break}{Last possible break point.}

\item{len}{Total number of observations.}

\item{SSR.data}{The matrix of recursive SSR values.}
}
\value{
List containing
\describe{
\item{SSR}{Optimal SSR value.}
\item{break_point}{The point of possible break.}
}
}
\description{
Procedure to minimize the SSR for 1 break point
}
\details{
See Carrion-i-Silvestre and Sansó (2006) for further details.
}
