% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.SSR.R
\name{segs.SSR.2.breaks}
\alias{segs.SSR.2.breaks}
\title{Procedure to minimize the SSR for 2 break points}
\usage{
segs.SSR.2.breaks(y, model)
}
\arguments{
\item{y}{(Tx1)-vector of time series}

\item{model}{\describe{
\item{1}{for the AA (without trend) model.}
\item{2}{for the AA (with trend) model.}
\item{3}{for the BB model.}
\item{4}{for the CC model.}
\item{5}{for the AC-CA model.}
}}
}
\value{
List containing
\describe{
\item{resid}{(Tx1) vector of estimated OLS residuals.}
\item{tb1}{The first break point.}
\item{tb2}{The second break point.}
}
}
\description{
Procedure to minimize the SSR for 2 break points
}
\details{
See Carrion-i-Silvestre and Sansó (2007) for further details.
}
