% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.matrix.R
\name{ssr.matrix}
\alias{ssr.matrix}
\title{Pre-calculate matrix of recursive SSR values.}
\usage{
ssr.matrix(y, x, width = 2)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Explanatory variables.}

\item{width}{Minimum spacing between the breaks.}
}
\value{
The matrix of recursive SSR values.
}
\description{
Pre-calculate matrix of recursive SSR values.
}
