% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSR.matrix.R
\name{SSR.matrix}
\alias{SSR.matrix}
\title{Pre-calculate matrix of recursive SSR values.}
\usage{
SSR.matrix(y, x, width = 2)
}
\arguments{
\item{y}{Dependent variable.}

\item{x}{Explanatory variables.}

\item{width}{Minimum spacing between the breaks.}
}
\value{
The matrix of recursive SSR values.
}
\description{
Pre-calculate matrix of recursive SSR values.
}
\details{
The function is not intended to be used directly so it's not exported.
}
