% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSR.recursive.R
\name{SSR.recursive}
\alias{SSR.recursive}
\title{Calculate SSR recursively}
\usage{
SSR.recursive(y, x, beg, end, width = 2)
}
\arguments{
\item{y}{(Tx1)-vector of the dependent variable.}

\item{x}{(Txk)-vector of the explanatory stochastic regressors.}

\item{beg}{The start of SSR calculating period.}

\item{end}{The end of SSR calculating period.}

\item{width}{Minimum spacing between the breaks.}
}
\value{
The vector of calculated recursive SSR.
}
\description{
Calculate SSR recursively
}
\details{
Based on Brown, Durbin and Evans (1975).
}
