% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STADF.test.R
\name{STADF.test}
\alias{STADF.test}
\alias{GSTADF.test}
\title{Supremum ADF tests with time transformation}
\usage{
STADF.test(
  y,
  trim = 0.01 + 1.8/sqrt(length(y)),
  const = FALSE,
  omega.est = TRUE,
  truncated = TRUE,
  is.reindex = TRUE,
  ksi.input = "auto",
  hc = 1,
  pc = 1,
  add.p.value = TRUE
)

GSTADF.test(
  y,
  trim = 0.01 + 1.8/sqrt(length(y)),
  const = FALSE,
  omega.est = TRUE,
  truncated = TRUE,
  is.reindex = TRUE,
  ksi.input = "auto",
  hc = 1,
  pc = 1,
  add.p.value = TRUE
)
}
\arguments{
\item{y}{A time series of interest.}

\item{trim}{A trimming parameter to determine the lower and upper bounds for
a possible break point.}

\item{const}{Whether the constant needs to be included.}

\item{omega.est}{Whether the variance of Nadaraya-Watson residuals should be
used.}

\item{truncated}{Whether the truncation of Nadaraya-Watson residuals is
needed.}

\item{is.reindex}{Whether the Cavaliere and Taylor (2008) time transformation
is needed.}

\item{ksi.input}{The value of the truncation parameter. Can be either \code{auto}
or the explicit numerical value. In the former case the numeric value is
estimated.}

\item{hc}{The scaling parameter for Nadaraya-Watson bandwidth.}

\item{pc}{The scaling parameter for the estimated truncation parameter value.}

\item{add.p.value}{Whether the p-value is to be returned. This argument is
needed to suppress the calculation of p-values during the precalculation of
tables needed for the p-values estimating.}
}
\value{
An object of type \code{sadf}. It's a list of:
\itemize{
\item \code{y},
\item \code{N}: Number of observations,
\item \code{trim},
\item \code{const},
\item \code{omega.est},
\item \code{truncated},
\item \code{is.reindex},
\item \code{new.index}: the vector of new indices,
\item \code{ksi.input},
\item \code{hc},
\item \code{h.est},
\item \code{u.hat},
\item \code{pc},
\item \code{w.sq},
\item \code{t.values}: vector of \eqn{t}-values,
\item the value of the corresponding test statistic,
\item \code{u.hat.truncated}: truncated residuals if truncation was asked for,
\item \code{ksi}, \code{sigma}: estimated values of the truncation parameter and resulting
s.e. if \code{ksi.input} equals \code{auto},
\item \code{eta.hat}: the values of reindexing function if reindexing was asked for,
\item \eqn{p}-value if it was asked for.
}
}
\description{
See \link{SADF.test}. Tests with time transformation are the modified versions of
the ordinary \code{SADF} and \code{GSADF} tests using Nadaraya-Watson residuals and
reindexing procedure by Cavaliere-Taylor (2008).
}
\references{
Cavaliere, Giuseppe, and A. M. Robert Taylor.
“Time-Transformed Unit Root Tests for Models with Non-Stationary Volatility.”
Journal of Time Series Analysis 29, no. 2 (March 2008): 300–330.
https://doi.org/10.1111/j.1467-9892.2007.00557.x.

Kurozumi, Eiji, Anton Skrobotov, and Alexey Tsarev.
“Time-Transformed Test for Bubbles under Non-Stationary Volatility.”
Journal of Financial Econometrics, April 23, 2022.
https://doi.org/10.1093/jjfinec/nbac004.
}
