% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STADF.test.R
\name{STADF.test}
\alias{STADF.test}
\title{SADF test based on NW estimation and reindexing.}
\usage{
STADF.test(
  y,
  r0 = 0.01 + 1.8/sqrt(length(y)),
  const = FALSE,
  omega.est = TRUE,
  h = "auto_CV",
  truncated = TRUE,
  is_reindex = TRUE,
  ksi.input = "auto",
  hc = 1,
  pc = 1,
  add.p.value = TRUE
)
}
\description{
SADF test based on NW estimation and reindexing.
}
