% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supSBADF.statistic.R
\name{supSBADF.statistic}
\alias{supSBADF.statistic}
\title{Calculate sup sign-based SADF statistic (HLZ, 2019)).}
\usage{
supSBADF.statistic(y, r0 = 0.01 + 1.8/sqrt(length(y)), generalized = FALSE)
}
\description{
Calculate sup sign-based SADF statistic (HLZ, 2019)).
}
