% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb.GSADF.test.R
\name{supSBADF.statistic}
\alias{supSBADF.statistic}
\title{Calculate superior sign-based SADF statistic.}
\usage{
supSBADF.statistic(y, trim = 0.01 + 1.8/sqrt(length(y)), generalized = FALSE)
}
\arguments{
\item{y}{The series of interest.}

\item{trim}{Trimming parameter to determine the lower and upper bounds.}

\item{generalized}{Whether to calculate generalized statistic value.}
}
\value{
A list of
\itemize{
\item \code{y},
\item \code{trim},
\item \code{C.t}: the cumulative sum of "signs" (1 or -1) of the first difference of
\code{y},
\item \code{SBADF.values}: series of sign-based ADF statistics,
\item \code{supSBADF.value}: the maximum of \code{SBADF.values}.
}
}
\description{
Calculate superior sign-based SADF statistic.
}
\references{
Harvey, David I., Stephen J. Leybourne, and Yang Zu.
“Sign-Based Unit Root Tests for Explosive Financial Bubbles in the Presence
of Deterministically Time-Varying Volatility.”
Econometric Theory 36, no. 1 (February 2020): 122–69.
https://doi.org/10.1017/S0266466619000057.
}
\keyword{internal}
